Move cypher projection to a new file
